% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abmodel-classdefs.r
\docType{class}
\name{abmodel-class}
\alias{abmodel-class}
\title{abmodel-class}
\description{
\strong{abmodel} is an S4 class that contains the ensemble model.
Besides the base learning algorithms--\code{base_models} --
\strong{abmodel} class contains information about the
dynamic selection method to apply in new data.
}
\section{Slots}{

\describe{
\item{\code{base_models}}{list comprising the tree learning models.}

\item{\code{form}}{formula}

\item{\code{data}}{training data for the tuning of the dynamic selection methods.}

\item{\code{dynamic_selection}}{Dynamic selection method to apply in new data. Character.
This value is predicted by the metalearning model.}
}}

\seealso{
\code{\link{autobagging}} function for the method os predicting the
best workflow.
}
