% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abmodel-classdefs.r
\name{abmodel}
\alias{abmodel}
\title{abmodel}
\usage{
abmodel(base_models, form, data, dynamic_selection)
}
\arguments{
\item{base_models}{adsd}

\item{form}{formula}

\item{data}{data}

\item{dynamic_selection}{asdada}
}
\description{
abmodel
}
