% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoBagging-pipeline.r, R/autoBagging.r
\docType{package}
\name{autoBagging}
\alias{autoBagging}
\alias{autoBagging}
\alias{autoBagging-package}
\title{autoBagging}
\usage{
autoBagging(form, data)
}
\arguments{
\item{form}{formula. Currently supporting only categorical target
variables (classification tasks)}

\item{data}{training dataset with a categorical target variable}
}
\value{
an \code{abmodel} class object
}
\description{
Learning to Rank Bagging Workflows with Metalearning

The autoBagging package provides three categories of important functions:
autoBagging, bar and baz.
}
\section{Foo functions}{

The autoBagging functions ...
}

\references{
Pinto, F. ...
}
\seealso{
\code{\link{bagging}} for the bagging pipeline with a specific
workflow; \code{\link{baggedtrees}} for the bagging implementation;
\code{\link{abmodel-class}} for the returning class object
}
