% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging.r
\name{baggedtrees}
\alias{baggedtrees}
\title{bagged trees models}
\usage{
baggedtrees(form, data, ntree = 100)
}
\arguments{
\item{form}{formula}

\item{data}{training data}

\item{ntree}{no of trees}
}
\description{
The standard resampling with replacement (bootstrap) is used
as sampling strategy.
}
\examples{
ensemble <- baggedtrees(Species ~., iris, ntree = 50)

}
