% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging.r
\name{bagging}
\alias{bagging}
\title{bagging method}
\usage{
bagging(form, data, ntrees, pruning, dselection, pruning_cp)
}
\arguments{
\item{form}{formula}

\item{data}{training data}

\item{ntrees}{ntrees}

\item{pruning}{model pruning method. A character vector. Currently, the
following methods are supported:
\describe{
   \item{mdsq}{sadsad}
   \item{bb}{boosting based..}
   \item{none}{no pruning}
}}

\item{dselection}{dynamic selection of the available models. Currently, the
following methods are supported:
\describe{
   \item{ola}{sadsad}
   \item{knora-e}{saddas..}
   \item{none}{no dynamic selection. Majority voting?}
}}

\item{pruning_cp}{The pruning cutpoint for the \code{pruning} method
picked.}
}
\description{
bagging method
}
\seealso{
\code{\link{baggedtrees}} for the implementation of the bagging model.
}
