% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarkers.r
\name{classmajority.landmarker}
\alias{classmajority.landmarker}
\title{classmajority.landmarker}
\usage{
classmajority.landmarker(dataset, data.char)
}
\arguments{
\item{dataset}{train data for the landmarker}

\item{data.char}{dc}
}
\description{
classmajority.landmarker
}
