% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarkers.r
\name{dstump.landmarker_d1.correlation}
\alias{dstump.landmarker_d1.correlation}
\title{dstump.landmarker_d1.correlation}
\usage{
dstump.landmarker_d1.correlation(dataset, data.char)
}
\arguments{
\item{dataset}{train data for the landmarker}

\item{data.char}{dc}
}
\description{
dstump.landmarker_d1.correlation
}
