% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarkers.r
\name{dstump.landmarker_d3.interinfo}
\alias{dstump.landmarker_d3.interinfo}
\title{dstump.landmarker_d3.interinfo}
\usage{
dstump.landmarker_d3.interinfo(dataset, data.char)
}
\arguments{
\item{dataset}{train data for the landmarker}

\item{data.char}{dc}
}
\description{
dstump.landmarker_d3.interinfo
}
