% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafeats-utils.r
\name{GetMeasure}
\alias{GetMeasure}
\title{Retrieve the value of a previously computed measure}
\usage{
GetMeasure(inDCName, inDCSet, component.name = "value")
}
\arguments{
\item{inDCName}{name of data characteristics}

\item{inDCSet}{set of data characteristics already computed}

\item{component.name}{name of component (e.g. time or value) to
retrieve; if NULL retrieve all}
}
\value{
simple or structured value
}
\description{
Retrieve the value of a previously computed measure
}
\note{
if measure is not available, stop execution with error
}
