% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicselection.r
\name{KNORA.E}
\alias{KNORA.E}
\title{pred asda}
\usage{
KNORA.E(form, mod, v.data, t.data, k = 5)
}
\arguments{
\item{form}{formula}

\item{mod}{lista modelos}

\item{v.data}{dados treino/val}

\item{t.data}{dados teste}

\item{k}{k do OLA as k instancias mais parecidas}
}
\description{
pred asda
}
