% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarkers.r
\name{lda.landmarker.correlation}
\alias{lda.landmarker.correlation}
\title{lda.landmarker.correlation}
\usage{
\method{lda}{landmarker.correlation}(dataset, data.char)
}
\arguments{
\item{dataset}{train data for the landmarker}

\item{data.char}{dc}
}
\description{
lda.landmarker.correlation
}
