% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.r
\name{mdsq}
\alias{mdsq}
\title{MDSQ}
\usage{
mdsq(form, preds, data, cutPoint)
}
\arguments{
\item{form}{formula}

\item{preds}{predictions in training data}

\item{data}{training data}

\item{cutPoint}{ratio of the total number of models to keepFORA}
}
\description{
MDSQ
}
