% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarkers.r
\name{nb.landmarker.mutual.information}
\alias{nb.landmarker.mutual.information}
\title{nb.landmarker.mutual.information}
\usage{
nb.landmarker.mutual.information(dataset, data.char)
}
\arguments{
\item{dataset}{train data for the landmarker}

\item{data.char}{dc}
}
\description{
nb.landmarker.mutual.information
}
