% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abmodel-methods.r
\docType{methods}
\name{predict,abmodel-method}
\alias{predict,abmodel-method}
\title{Predicting on new data with a \strong{autoBagging} model}
\usage{
\S4method{predict}{abmodel}(object, newdata)
}
\arguments{
\item{object}{A \strong{abmodel-class} object.}

\item{newdata}{New data to predict using \code{autobagging} model object}
}
\value{
predictions made by an \code{autobagging} model.
}
\description{
This is a \code{predict} method for predicting new data points using a
\code{tseModel} class object - refering to an ensemble for time series
forecasting.
}
\seealso{
\code{\link{abmodel-class}} for details about the ensemble model;
}
