% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicselection.r
\name{predict.KNORA.E}
\alias{predict.KNORA.E}
\title{pred asda}
\usage{
predict.KNORA.E(form, mod, v.data, t.data, k = 5)
}
\arguments{
\item{form}{formula}

\item{mod}{lista modelos}

\item{v.data}{dados treino/val}

\item{t.data}{dados teste}

\item{k}{k do OLA as k instancias mais parecidas}
}
\description{
pred asda
}
