% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafeats-utils.r
\name{ReadDF}
\alias{ReadDF}
\title{FUNCTION TO TRANSFORM DATA FRAME INTO LIST WITH GSI REQUIREMENTS}
\usage{
ReadDF(dat)
}
\arguments{
\item{dat}{data frame}
}
\value{
a list containing components that describe
the names (see ReadtAttrsInfo) and the data (see ReadData) files

THIS FUNCTION HAS TO BE BASED IN READATTRSINFO AND READDATA
}
\description{
FUNCTION TO TRANSFORM DATA FRAME INTO LIST WITH GSI REQUIREMENTS
}
