% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafeats-utils.r
\name{SymbAttrs}
\alias{SymbAttrs}
\title{Retrieve names of symbolic attributes (not including the target)}
\usage{
SymbAttrs(dataset)
}
\arguments{
\item{dataset}{structure describing the data set, according
to \code{read_data.R}}
}
\value{
list of strings
}
\description{
Retrieve names of symbolic attributes (not including the target)
}
\seealso{
read_data.R
}
