% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abmodel-classdefs.r
\docType{class}
\name{abmodel-class}
\alias{abmodel-class}
\title{abmodel-class}
\description{
\strong{abmodel} is an S4 class that contains the ensemble model.
Besides the base learning algorithms--\code{base_models} --
\strong{abmodel} class contains information about the
dynamic selection method to apply in new data.
}
\section{Slots}{

\describe{
\item{\code{base_models}}{a list of decision tree classifiers}

\item{\code{form}}{formula}

\item{\code{data}}{dataset used to train \code{base_models}}

\item{\code{dynamic_selection}}{the dynamic selection/combination method
to use to aggregate predictions. If \code{none}, majority vote is used.}
}}

\seealso{
\code{\link{autoBagging}} function for the
method of automatic predicting of the best workflows.
}
