% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abmodel-classdefs.r
\name{abmodel}
\alias{abmodel}
\title{abmodel}
\usage{
abmodel(base_models, form, data, dynamic_selection)
}
\arguments{
\item{base_models}{a list of decision tree classifiers}

\item{form}{formula}

\item{data}{dataset used to train \code{base_models}}

\item{dynamic_selection}{the dynamic selection/combination method
to use to aggregate predictions. If \code{none}, majority vote is used.}
}
\description{
abmodel
}
