% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autobagging-pipeline.r, R/autobagging.r
\docType{package}
\name{autobagging}
\alias{autobagging}
\alias{autobagging}
\alias{autobagging-package}
\title{autobagging}
\usage{
autobagging(form, data)
}
\arguments{
\item{form}{formula. Currently supporting only categorical target
variables (classification tasks)}

\item{data}{training dataset with a categorical target variable}
}
\value{
an \code{abmodel} class object
}
\description{
Learning to Rank Bagging Workflows with Metalearning

The autobagging package provides three categories of important functions:
autobagging, bar and baz.
}
\section{Foo functions}{

The autobagging functions ...
}

\references{
Pinto, F. ...
}
\seealso{
\code{\link{bagging}} for the bagging pipeline with a specific
workflow; \code{\link{baggedtrees}} for the bagging implementation;
\code{\link{abmodel-class}} for the returning class object
}
