% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicselection.r
\name{OLA}
\alias{OLA}
\title{Overall Local Accuracy}
\usage{
OLA(form, mod, v.data, t.data, k = 5)
}
\arguments{
\item{form}{formula}

\item{mod}{a list comprising the individual models}

\item{v.data}{validation data}

\item{t.data}{test data, with the instances to predict}

\item{k}{the number of nearest neighbors. Defaults to 5.}
}
\description{
A dynamic selection method
}
