% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abmodel-methods.r
\docType{methods}
\name{predict,abmodel-method}
\alias{predict,abmodel-method}
\title{Predicting on new data with a \strong{abmodel} model}
\usage{
\S4method{predict}{abmodel}(object, newdata)
}
\arguments{
\item{object}{A \strong{abmodel-class} object.}

\item{newdata}{New data to predict using an \code{abmodel} object}
}
\value{
predictions produced by an \code{abmodel} model.
}
\description{
This is a \code{predict} method for predicting new data points using a
\code{abmodel} class object - refering to an ensemble
of bagged trees
}
\seealso{
\code{\link{abmodel-class}} for details about the bagging model;
}
